migrations adjusted & order service changed & repos changed

+++ internal/pkg/repository/schema/20250419164922_create_withdrawals_table.sql
@@ -3,7 +3,7 @@
 CREATE TABLE withdrawals (
     id SERIAL PRIMARY KEY,
     user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-    order_id INTEGER NOT NULL,
+    order_id BIGINT NOT NULL,
     sum NUMERIC(20, 2) NOT NULL,
     processed_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
